requirements.txt
sh
pymongo
pika
pyyaml
termcolor
